requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
openai
pymupdf
python-multipart
python-dotenv
sse-starlette
